+++ cmake/ExternalDependencies.cmake
@@ -70,7 +70,7 @@ set(QDMI_MINIMUM_VERSION 1.4.0
         CACHE STRING "Minimum QDMI version")
 set(QDMI_VERSION 1.4.0
         CACHE STRING "QDMI version")
-set(QDMI_REV "dcb57425fe650867344e5989eecc77d42231c3a4"
+set(QDMI_REV "d6ce0da02cba1a3b903f0b046ef23af6f00edc6a"
         CACHE STRING "QDMI identifier (tag, branch or commit hash)")
 set(QDMI_REPO_OWNER "Munich-Quantum-Software-Stack"
         CACHE STRING "QDMI repository owner (change when using a fork)")